requirements.txt
pdfminer.six
python-docx
spacy
nltk
pandas
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
